PC user (#2103)
* fix:pc user camp and learner pagination

* fix: pc user isse resolve

* fix:sonar issue

* fix:sonar
