+++ apps/front-end/src/pages/PCusers/Learner/LearnerProfileView.js
@@ -1,17 +1,18 @@
-import React, { useEffect } from "react";
-import { useParams, useNavigate, useLocation } from "react-router-dom";
-import { HStack, VStack, Divider, Stack } from "native-base";
 import {
   FrontEndTypo,
   IconByName,
-  PCusers_layout as Layout,
   ImageView,
+  PCusers_layout as Layout,
   PcuserService,
 } from "@shiksha/common-lib";
-import Clipboard from "component/Clipboard";
+import Menu from "component/Beneficiary/Menu";
 import { ChipStatus } from "component/BeneficiaryStatus";
-import { useTranslation } from "react-i18next";
+import Clipboard from "component/Clipboard";
+import { Stack, VStack } from "native-base";
 import PropTypes from "prop-types";
+import React, { useEffect } from "react";
+import { useTranslation } from "react-i18next";
+import { useLocation, useNavigate, useParams } from "react-router-dom";
 
 export default function LearnerProfileView({ userTokenInfo }) {
   const navigate = useNavigate();
@@ -182,50 +183,4 @@ export default function LearnerProfileView({ userTokenInfo }) {
 
 LearnerProfileView.PropTypes = {
   userTokenInfo: PropTypes.any,
-};
-
-const Menu = ({ menus }) => {
-  const { t } = useTranslation();
-
-  return (
-    <VStack
-      shadow={"LearnerProfileViewShadow"}
-      bg="boxBackgroundColour.100"
-      borderColor="garyTitleCardBorder"
-      borderRadius="5px"
-      borderWidth="1px"
-      px="4"
-      p="2"
-      pb="3"
-      divider={
-        <Divider orientation="horizontal" bg="btnGray.100" thickness="1" />
-      }
-    >
-      {menus?.map((menu) => (
-        <HStack
-          key={menu}
-          alignItems="center"
-          justifyContent="space-between"
-          p="2"
-          pr="0"
-        >
-          <FrontEndTypo.H3 color="floatingLabelColor.500" fontWeight={"600"}>
-            {t(menu?.title)}
-          </FrontEndTypo.H3>
-
-          <IconByName
-            name="ArrowRightSLineIcon"
-            onPress={menu?.onPress || (() => console.log("not onpress"))}
-            color="floatingLabelColor.500"
-            _icon={{ size: "20" }}
-            {...(menu?._icon || {})}
-          />
-        </HStack>
-      ))}
-    </VStack>
-  );
-};
-
-Menu.PropTypes = {
-  menus: PropTypes.array,
 };
