+++ lib/common-lib/src/components/Breadcrumb.js
@@ -36,11 +36,7 @@ export default function Breadcrumb({
                 {item?.label ? item?.label : item?.title ? item?.title : ''}
               </Text>
             )}
-            {data.length > index + 1 ? (
-              drawer || <Text>/</Text>
-            ) : (
-              <React.Fragment />
-            )}
+            {data.length > index + 1 && (drawer || <Text>/</Text>)}
           </HStack>
         ))}
       </HStack>
