+++ lib/common-lib/src/components/PCusers_layout/Layout.js
@@ -23,7 +23,7 @@ export default function App({ _footer, _drawer, user, facilitator, ...props }) {
           },
           {
             title: 'PRERAKS',
-            route: '/prerak/PrerakList',
+            route: '/preraks',
             icon: 'GraduationCap'
           },
           {
